Updated prettier print width and ran prettier write.

+++ app/hooks/useBookmarks.ts
@@ -5,30 +5,22 @@ import { BookmarksAPI } from '@app/api/bookmarks';
 import { IBookmark } from '@app/types';
 
 export const useBookmarks = () => {
-  const result = useQuery<IBookmark[], string>(
-    ['bookmarks'],
-    BookmarksAPI.fetchAllBookmarks,
-    {
-      onError: error => console.error(error),
-    },
-  );
+  const result = useQuery<IBookmark[], string>(['bookmarks'], BookmarksAPI.fetchAllBookmarks, {
+    onError: error => console.error(error),
+  });
 
   const deleteBookmark = async (id: string) => {
-    Alert.alert(
-      'Delete bookmark?',
-      'This will permanently delete the bookmark and remove it from any collections.',
-      [
-        {
-          text: 'Delete',
-          onPress: async () => {
-            await BookmarksAPI.deleteBookmark(id);
-            result.refetch();
-          },
-          style: 'destructive',
+    Alert.alert('Delete bookmark?', 'This will permanently delete the bookmark and remove it from any collections.', [
+      {
+        text: 'Delete',
+        onPress: async () => {
+          await BookmarksAPI.deleteBookmark(id);
+          result.refetch();
         },
-        { text: 'Cancel', style: 'cancel' },
-      ],
-    );
+        style: 'destructive',
+      },
+      { text: 'Cancel', style: 'cancel' },
+    ]);
   };
 
   return { ...result, deleteBookmark };
